+++ lib/app_widgets.dart
@@ -3,3 +3,4 @@ export 'package:vamstreet_components/src/features/widgets/buttons/adaptive_prima
 export 'package:vamstreet_components/src/features/widgets/buttons/adaptive_secondary_button/adaptive_secondary_button.dart';
 export 'package:vamstreet_components/src/features/widgets/buttons/adaptive_text_button/adaptive_text_button.dart';
 export 'package:vamstreet_components/src/features/widgets/checkboxes/adaptive_checkbox/adaptive_checkbox_checkbox.dart';
+export 'package:vamstreet_components/src/features/widgets/text_inputs/adaptive_text_field/adaptive_text_field_widget.dart';