+++ lib/src/core/bloc/button_color_cubit.dart
@@ -1,0 +1,8 @@
+import 'package:fluent_ui/fluent_ui.dart' show Color;
+import 'package:flutter_bloc/flutter_bloc.dart' show Cubit;
+
+class ButtonColorCubit extends Cubit<Color> {
+  ButtonColorCubit(super.initialState);
+
+  void updateColor(Color newColor) => emit(newColor);
+}