+++ lib/src/features/widgets/buttons/adaptive_text_button/adaptive_text_button.dart
@@ -4,8 +4,9 @@ import 'package:vamstreet_components/src/core/enum/padding_size_enum.dart';
 import 'package:vamstreet_components/src/core/enum/request_state_enum.dart';
 import 'package:vamstreet_components/src/core/widgets/responsive_width_container.dart'
     show ResponsiveWidthContainer;
+import 'package:vamstreet_components/src/features/widgets/buttons/adaptive_text_button/text_button_widget.dart'
+    show TextButtonWidget;
 import 'package:vamstreet_components/src/features/widgets/buttons/core/enums/button_type_enum.dart';
-import 'package:vamstreet_components/src/features/widgets/buttons/core/common/widgets/adaptive_button_content_widget.dart';
 
 /// ## [AdaptiveTextButton] Class Documentation
 ///
@@ -185,41 +186,19 @@ class AdaptiveTextButton extends StatelessWidget {
         constraints: BoxConstraints(minHeight: minHeight, maxWidth: maxWidth),
         child: MouseRegion(
           cursor: SystemMouseCursors.click,
-          child: Button(
-            style: ButtonStyle(
-              padding: WidgetStateProperty.all(
-                EdgeInsets.symmetric(horizontal: paddingSize.horizontalSize),
-              ),
-              backgroundColor: WidgetStateProperty.all(Colors.transparent),
-              foregroundColor: WidgetStateProperty.all<Color>(buttonColor),
-              elevation: WidgetStateProperty.all(0),
-              shape: WidgetStateProperty.all(
-                RoundedRectangleBorder(
-                  borderRadius: BorderRadius.circular(borderRadius),
-                ),
-              ),
-            ),
-            onPressed:
-                (requestState == RequestStateEnum.loading) ? null : onPressed,
-            child: Padding(
-              padding: EdgeInsets.symmetric(
-                horizontal: paddingSize.horizontalSize,
-              ),
-              child: AdaptiveButtonContent(
-                textOverflow: textOverflow,
-                requestState: requestState,
-                suffixIcon: suffixIcon,
-                prefixIcon: prefixIcon,
-                title: title,
-                buttonType: buttonType,
-                icon: baseIcon,
-                paddingSize: paddingSize,
-                contentColorMode:
-                    isDefaultTextButton
-                        ? ContentColorMode.defaultMode
-                        : ContentColorMode.accentMode,
-              ),
-            ),
+          child: TextButtonWidget(
+            paddingSize: paddingSize,
+            buttonColor: buttonColor,
+            borderRadius: borderRadius,
+            requestState: requestState,
+            onPressed: onPressed,
+            textOverflow: textOverflow,
+            suffixIcon: suffixIcon,
+            prefixIcon: prefixIcon,
+            title: title,
+            buttonType: buttonType,
+            baseIcon: baseIcon,
+            isDefaultTextButton: isDefaultTextButton,
           ),
         ),
       ),