+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 - [Update]: fix import widget for access outer `src` directory  [#17](https://github.com/adelan-tejarat/vamstreet-components/pull/19)
 
+- [Update]: fix import `text_field_sizes` for access outer `src` directory   [#17](https://github.com/adelan-tejarat/vamstreet-components/pull/21)
+
+- [Update]: match dark accent color to light mode value [#17](https://github.com/adelan-tejarat/vamstreet-components/pull/21)
+
 # 1.8.0+10
 - [Refactor]:  Create Checkbox Component [#15](https://github.com/adelan-tejarat/vamstreet-components/pull/16)
   - feat(checkboxes): add adaptive checkbox widget